+++ src/clojure_lsp/feature/code_actions.clj
@@ -12,11 +12,11 @@
       CodeActionKind)))
 
 (defn ^:private find-missing-require [uri diagnostic]
-  (let [{{:keys [line character] :as position} :start} (:range diagnostic)
-        diagnostic-zloc (parser/cursor-zloc uri line character)]
-    (when-let [missing-require (r.transform/find-missing-require diagnostic-zloc)]
-      {:missing-require missing-require
-       :position        position})))
+  (let [{{:keys [line character] :as position} :start} (:range diagnostic)]
+    (when-let [diagnostic-zloc (parser/safe-cursor-loc uri line character)]
+      (when-let [missing-require (r.transform/find-missing-require diagnostic-zloc)]
+        {:missing-require missing-require
+         :position        position}))))
 
 (defn ^:private find-missing-requires [uri diagnostics]
   (let [unresolved-ns-diags (filter #(= "unresolved-namespace" (:code %)) diagnostics)
@@ -31,11 +31,11 @@
          (remove nil?))))
 
 (defn ^:private find-missing-import [uri diagnostic]
-  (let [{{:keys [line character] :as position} :start} (:range diagnostic)
-        diagnostic-zloc (parser/cursor-zloc uri line character)]
-    (when-let [missing-import (r.transform/find-missing-import diagnostic-zloc)]
-      {:missing-import missing-import
-       :position       position})))
+  (let [{{:keys [line character] :as position} :start} (:range diagnostic)]
+    (when-let [diagnostic-zloc (parser/safe-cursor-loc uri line character)]
+      (when-let [missing-import (r.transform/find-missing-import diagnostic-zloc)]
+        {:missing-import missing-import
+         :position       position}))))
 
 (defn ^:private find-missing-imports [uri diagnostics]
   (let [unresolved-ns-diags (filter #(= "unresolved-namespace" (:code %)) diagnostics)