+++ src/clojure_lsp/handlers.clj
@@ -326,12 +326,12 @@
         character (-> range :start :character)
         row (inc line)
         col (inc character)
-        zloc (parser/cursor-zloc textDocument line character)
+        zloc (parser/safe-cursor-loc textDocument line character)
         client-capabilities (get db :client-capabilities)]
     (f.code-actions/all zloc textDocument row col diagnostics client-capabilities)))
 
 (defn resolve-code-action [{{:keys [uri line character]} :data :as action}]
-  (let [zloc (parser/cursor-zloc uri line character)]
+  (let [zloc (parser/safe-cursor-loc uri line character)]
     (f.code-actions/resolve-code-action action zloc)))
 
 (defn code-lens